Fix unwarranted 'UnsupportedLockDependency' errors
Ultimately we would want to raise an exception when the dependency
definition in the lockfile is unsupported, but it would be raised in
many unwarranted conditions as well. Just listing the environments would
trigger the exception.

So we print an error message instead.

GitHub: closes https://github.com/sinoroc/tox-poetry-dev-dependencies/issues/54